# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Patient Medicine Tracker (PMM) API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://pmm.oloma.dev/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/modules/findAll": {
            "get": {
                "tags": [
                    "Modules"
                ],
                "summary": "Find all modules",
                "operationId": "modules_findAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModulesFindAll"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
        "/modules/update/{moduleId}": {
            "put": {
                "tags": [
                    "Modules"
                ],
                "summary": "Update module",
                "operationId": "modules_update",
                "parameters": [
                    {
                        "name": "moduleId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update role",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModuleSave"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModuleUpdateResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
...
        "/users/delete/{userId}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user",
                "operationId": "users_delete",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "user uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
        "/users/myAccount/updatePassword": {
            "put": {
                "tags": [
                    "Users My Account"
                ],
                "summary": "Update password",
                "operationId": "usersMyAccount_updatePassword",
                "requestBody": {
                    "description": "Update Password",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MyAccountUpdatePassword"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
        "/users/update/{userId}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "operationId": "users_update",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update user",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSave"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AvatarObject": {
                "properties": {
                    "image": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ModuleSave": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "version": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "ModuleUpdateResponse": {
                "properties": {
                    "oldRecord": {
                        "$ref": "#/components/schemas/ModuleUpdateResponseObject"
                    }
                },
                "type": "object"
            },
            "ModuleUpdateResponseObject": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "version": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "ModulesFindAll": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "version": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "MyAccountUpdatePassword": {
                "properties": {
                    "oldPassword": {
                        "type": "string"
                    },
                    "newPassword": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UserSave": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "avatar": {
                        "$ref": "#/components/schemas/AvatarObject"
                    },
                    "isActive": {
                        "type": "integer"
                    },
                    "isEmailActivated": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}